# license-service/api/swagger/v1/license.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "license.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/equipments/types/{equip_type}/{equip_id}/metric/types/{metric_type}/{metric_name}": {
      "post": {
        "operationId": "LicensesForEquipAndMetric",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LicensesForEquipAndMetricResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "equip_type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "equip_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "metric_type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "metric_name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LicensesForEquipAndMetricRequest"
            }
          }
        ],
        "tags": [
          "LicenseService"
        ]
      }
    },
    "/api/v1/licenses/applications/{app_id}/products/{prod_id}": {
      "get": {
        "operationId": "ListAcqRightsForApplicationsProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAcqRightsForApplicationsProductResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "prod_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LicenseService"
        ]
      }
    },
    "/api/v1/product/{swid_tag}/acquiredrights": {
      "get": {
        "operationId": "ListAcqRightsForProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAcquiredRightsForProductResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "swid_tag",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LicenseService"
        ]
      }
    },
    "/api/v1/product/{swid_tag}/metric/{metric_name}": {
      "post": {
        "operationId": "ProductLicensesForMetric",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ProductLicensesForMetricResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "swid_tag",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "metric_name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ProductLicensesForMetricRequest"
            }
          }
        ],
        "tags": [
          "LicenseService"
        ]
      }
    },
    "/api/v1/products/aggregations": {
      "post": {
        "operationId": "CreateProductAggregation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ProductAggregation"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ProductAggregation"
            }
          }
        ],
        "tags": [
          "LicenseService"
        ]
      }
    },
    "/api/v1/products/aggregations/productview/{ID}/acquiredrights": {
      "get": {
        "operationId": "ListAcqRightsForProductAggregation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAcqRightsForProductAggregationResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LicenseService"
        ]
      }
    }
  },
  "definitions": {
    "v1Attribute": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "data_type": {
          "$ref": "#/definitions/v1DataTypes"
        },
        "primary_key": {
          "type": "boolean",
          "format": "boolean"
        },
        "displayed": {
          "type": "boolean",
          "format": "boolean"
        },
        "searchable": {
          "type": "boolean",
          "format": "boolean"
        },
        "parent_identifier": {
          "type": "boolean",
          "format": "boolean"
        },
        "mapped_to": {
          "type": "string"
        },
        "simulated": {
          "type": "boolean",
          "format": "boolean"
        },
        "int_val": {
          "type": "integer",
          "format": "int32"
        },
        "float_val": {
          "type": "number",
          "format": "float"
        },
        "string_val": {
          "type": "string"
        },
        "int_val_old": {
          "type": "integer",
          "format": "int32"
        },
        "float_val_old": {
          "type": "number",
          "format": "float"
        },
        "string_val_old": {
          "type": "string"
        }
      }
    },
    "v1DataTypes": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "STRING",
        "INT",
        "FLOAT"
      ],
      "default": "UNKNOWN"
    },
    "v1LicensesForEquipAndMetricRequest": {
      "type": "object",
      "properties": {
        "equip_type": {
          "type": "string"
        },
        "equip_id": {
          "type": "string"
        },
        "metric_type": {
          "type": "string"
        },
        "metric_name": {
          "type": "string"
        },
        "attributes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Attribute"
          }
        },
        "scope": {
          "type": "string"
        }
      }
    },
    "v1LicensesForEquipAndMetricResponse": {
      "type": "object",
      "properties": {
        "licenses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ProductLicenseForEquipAndMetric"
          }
        }
      }
    },
    "v1ListAcqRightsForApplicationsProductResponse": {
      "type": "object",
      "properties": {
        "acq_rights": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ProductAcquiredRights"
          }
        }
      }
    },
    "v1ListAcqRightsForProductAggregationResponse": {
      "type": "object",
      "properties": {
        "acq_rights": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ProductAcquiredRights"
          }
        }
      }
    },
    "v1ListAcquiredRightsForProductResponse": {
      "type": "object",
      "properties": {
        "acq_rights": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ProductAcquiredRights"
          }
        }
      }
    },
    "v1Product": {
      "type": "object",
      "properties": {
        "swidTag": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "editor": {
          "type": "string"
        },
        "metric": {
          "type": "string"
        },
        "numCptLicences": {
          "type": "integer",
          "format": "int32"
        },
        "numAcqLicences": {
          "type": "integer",
          "format": "int32"
        },
        "totalCost": {
          "type": "number",
          "format": "double"
        },
        "deltaNumber": {
          "type": "integer",
          "format": "int32"
        },
        "deltaCost": {
          "type": "number",
          "format": "double"
        },
        "numOfApplications": {
          "type": "integer",
          "format": "int32"
        },
        "numofEquipments": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1ProductAcquiredRights": {
      "type": "object",
      "properties": {
        "SKU": {
          "type": "string"
        },
        "swidTag": {
          "type": "string"
        },
        "metric": {
          "type": "string"
        },
        "numCptLicences": {
          "type": "integer",
          "format": "int32"
        },
        "numAcqLicences": {
          "type": "integer",
          "format": "int32"
        },
        "totalCost": {
          "type": "number",
          "format": "double"
        },
        "deltaNumber": {
          "type": "integer",
          "format": "int32"
        },
        "deltaCost": {
          "type": "number",
          "format": "double"
        },
        "avgUnitPrice": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "v1ProductAggregation": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "editor": {
          "type": "string"
        },
        "product_names": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "metric": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "products_full": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Product"
          }
        }
      }
    },
    "v1ProductLicenseForEquipAndMetric": {
      "type": "object",
      "properties": {
        "metric_name": {
          "type": "string"
        },
        "old_licences": {
          "type": "string",
          "format": "int64"
        },
        "new_licenses": {
          "type": "string",
          "format": "int64"
        },
        "delta": {
          "type": "string",
          "format": "int64"
        },
        "product": {
          "$ref": "#/definitions/v1Product"
        }
      }
    },
    "v1ProductLicensesForMetricRequest": {
      "type": "object",
      "properties": {
        "swid_tag": {
          "type": "string"
        },
        "metric_name": {
          "type": "string"
        },
        "unit_cost": {
          "type": "number",
          "format": "double"
        },
        "scope": {
          "type": "string"
        }
      }
    },
    "v1ProductLicensesForMetricResponse": {
      "type": "object",
      "properties": {
        "numCptLicences": {
          "type": "string",
          "format": "uint64"
        },
        "total_cost": {
          "type": "number",
          "format": "double"
        },
        "metric_name": {
          "type": "string"
        }
      }
    }
  }
}